Added option for natural arrow keys
The natural order is used for the arrow keys. Arrow up select the upper
slider value for vertical sliders, arrow right the righter slider value
for a horizontal slider - no matter if the slider was reversed or not.
By default the arrow keys are oriented by arrow up/right to the higher
slider value, arrow down/left to the lower slider value.

Tests are provided.