Clean mousemove and mouseup events
If you press one mouse button and then another without releasing the first one, two mouseup (and mousemove) events are added to document. When you release both buttons, document.removeEventListener is trying to remove this.mouseup from document mouseup events array, but it can find just the newer one and not the old one, so the slider remains in a pressed state.