Prevent bootstrap modal hiding
Issue #339: Modal dialog closes when dragging and releasing mouse button outside dialog.
Bootstrap modal needs a mousedown event to check the modal needs to be closed. When you release the mouse, if mouse was on modal, there will be no exit. But in our situation bootstrap-slider prevents bubbling mousedown event to modal, so it closes when you finished mouse movement outside modal.